Create ticTacToe
I deviated from the original code and I wanted to add something in that will check for a winner. Checking back to chapter 5 using the coordinates of the Grid, I utilized a portion of the for loop code in hopes of checking if there were cells (key values) that had 3 of the same characters in a row vertically, diagonal, or horizontally. 

I am not exactly sure how to go about this. I was thinking maybe a for loop that would count a character(character being "X' or 'O') and if it is not in sequence, have the code end and go to the next vertical, diagonal, or horizontal. So, a for loop that would iterate over the 'theBoard' values. Completely lost at this point and any additional assistance would be fantastic.